Resolução de erros de tipagem e bugs nos componentes de Subjects e Exams #158
