+++ src/functions/Exams/ActivateExamFunction.ts
