+++ src/functions/Exams/CreateExam.ts
@@ -1,9 +1,21 @@
 import { doc, updateDoc } from 'firebase/firestore';
 import { auth, db } from '../../../config/firebase';
 
-import { IExam } from '../../components/Exams/examInterfaces';
+import { Timestamp } from 'firebase/firestore';
 
-const CreateExamFunction = async (subjectId: string, exam: IExam) => {
+export interface ICreateExam {
+  code: string;
+  score: string;
+  date: Timestamp | null;
+  room: string;
+  status: string;
+  id: string;
+  time: Timestamp | null;
+  codeSubject: string;
+  subjectID: string;
+}
+
+const CreateExamFunction = async (subjectId: string, exam: ICreateExam) => {
   if (!auth.currentUser) return;
   try {
     const examID = Math.random().toString(36).substring(7);
