+++ src/functions/Exams/DeleteExam.ts
